refactor(bills): Config structure, backward compatible - Move bill types to constants/bills/types - Add a union type for the three different bill types which enforces types based on the billType param
