--- src/constants/bills/billsLaunch.ts
@@ -1,3 +1,0 @@
-import { BillsLaunchConfig } from '../../types'
-
-export const billsLaunch: BillsLaunchConfig[] = []
