+++ src/constants/bills/index.ts
@@ -1,6 +1,2 @@
-import { bills } from './bills'
-import { billsLaunch } from './billsLaunch'
-
-const allBills = [...bills, ...billsLaunch]
-
-export { bills, billsLaunch, allBills }
+export * from './bills'
+export * from './types'
